+++ src/main/java/nextstep/member/repository/MemberRepository.java
@@ -1,5 +1,6 @@
 package nextstep.member.repository;
 
+import nextstep.member.domain.Member;
 import nextstep.member.entity.MemberEntity;
 
 import java.util.Optional;
@@ -8,7 +9,7 @@ public interface MemberRepository {
 
     Long save(MemberEntity memberEntity);
 
-    Optional<MemberEntity> findById(Long id);
+    Optional<Member> findById(Long id);
 
-    Optional<MemberEntity> findByUsername(String username);
+    Optional<Member> findByUsername(String username);
 }
